requirements.txt
discord.py
groq
aiohttp